Simplified code.
