+++ src/array/boolean/ffi.rs
@@ -1,5 +1,6 @@
 use crate::{
     array::{FromFfi, ToFfi},
+    datatypes::DataType,
     ffi,
 };
 
@@ -22,6 +23,8 @@ unsafe impl ToFfi for BooleanArray {
 
 unsafe impl<A: ffi::ArrowArrayRef> FromFfi<A> for BooleanArray {
     fn try_from_ffi(array: A) -> Result<Self> {
+        let data_type = array.field()?.data_type().clone();
+        assert_eq!(data_type, DataType::Boolean);
         let length = array.array().len();
         let offset = array.array().offset();
         let mut validity = unsafe { array.validity() }?;
